+++ hpsspy/data/tests/test_data.py
@@ -16,6 +16,13 @@ class TestData(unittest.TestCase):
         for release in ['dr{0:d}'.format(k) for k in range(8,13)]:
             self.assertIn(release,hpss_map,"Release {0} is not in sdss.json!".format(release))
         t.close()
+    def test_desi(self):
+        self.assertTrue(resource_exists('hpsspy.data','desi.json'),"Could not find desi.json!")
+        t = resource_stream('hpsspy.data','desi.json')
+        hpss_map = json.load(t)
+        for release in ('datachallenge','imaging','mocks','release','spectro','target'):
+            self.assertIn(release,hpss_map,"Release {0} is not in sdss.json!".format(release))
+        t.close()
 #
 if __name__ == '__main__':
     unittest.main()