+++ doc/changes.rst
@@ -7,6 +7,7 @@ Release Notes
 
 * Moved configuration items to JSON files.
 * Started adding support for DESI.
+* Add tests to util subpackage.
 
 0.1.0 (2015-03-25)
 ------------------